Add contexts Clean up symbols Remove dispatchers and actions
Problem: parameters don't parse properly when constructing a token block
TODO: substitute parameters for corresponding tokens.  That should make it parse.

Problem: how am I going to do subclassing?
I think the parser generator has to know about it.  It can't just be productions
of the form SUPER -> SUB because in the current semantics, that implies a wrapper
object or a tag change, and neither is acceptable because subsequent parsing won't
recognize it as a SUB anymore.


git-svn-id: file:///svn/ellesmere@73 8a67d1f4-d520-4d41-b900-46b8824da3b7