Make sure array indices map to different pseudo-SymbolIndex values from real symbols. That way a single object can have both fields and elements.
Should be useful to build the graph representing the symbol DAG that I'm 
planning to use for expressing subtype relations etc.  I spent a week 
thinking of making integers symbols, but that raised a couple of 
problems:

1. Now there would suddenly be an infinite number of symbols.  That 
makes the notion of a symbol table somewhat meaningless.  I could fake 
it I suppose, but there seems to be a measure of sanity in the notion 
that you can ask a symbol table how many symbols there are!

2. I don't want integers to be all that special.  Yes, small integers 
are treated specially in the implementation so ordinary math doesn't 
require object allocations, but that performance optimization doesn't 
mean I want to elevate integers to a special status by incorporating 
them into the set of symbols.  What next?  What about floating-point 
numbers?  If I'm ever going to be able to handle terminal-case 
statements like def fib(1) as { 0 }, then I don't want to do it by 
treating ints as symbols.  I'll have to be able to parse actual objects 
by some means if I'm ever going to support user types. Whatever means I 
pick, it will presumably work for integers, so there's no need to treat 
them so specially.


git-svn-id: file:///svn/ellesmere@123 8a67d1f4-d520-4d41-b900-46b8824da3b7