NON FUNCTIONAL VERSION
I tried extra augmenting, but I still have a problem when parsing (not running) some code:
I may not know which subtag will be found, so the parsing process might actually need to
deal with the token for a supertag.

This is an interesting phenomenon.  It is directly analogous to the situation in a
statically typed language where a feature common to multiple subclasses is not declared
in the superclass.  Just because DumpTruck has wheels() and FireTruck has wheels()
doesn't mean we know all possible instances of Truck will have wheels().  So for a while,
I thought I had just discovered that programs in my language must declare the equivalent
of an abstract method for the supertag.

However, this can add a lot of verbiage just to satisfy the parser, and the fact
remains that dynamically-typed languages do not need this sort of verbiage.  I could
declare DumpTruck.wheels() and FireTruck.wheels() in python, and then call truck.wheels(),
and expect that to work without declaring wheels() in any common superclass.  Therefore,
if Ellesmere is to be a contender for a universal programming language, I can't require
this kind of declaration without dooming it to be less powerful than Python in some
situations.

I think the way to deal with that is to augment Vijay's algorithm so it goes in both
directions.  I'm planning to augment the automata to add additional reduce actions, but
I'll use negative numbers instead of positive numbers.  Then, when doing a reduce, if
it's a positive integer N, reduce by item N; else, if it's a negative integer -N,
reduce by N unless we're executing the code (rather than recording it or otherwise
just parsing it), report the equivalent of an abstract method error.  If the tag
in question has multiple subtags with reduce actions, their LHS tokens must match,
as must the lengths of their RHSes, or else report an error at augment time.



git-svn-id: file:///svn/ellesmere@134 8a67d1f4-d520-4d41-b900-46b8824da3b7