I had been working toward drastically simplifying the parsing of TOKEN_BLOCKs so they are not required to parse the contents faithfully.  The digression model is too neat and simple to be hamstrung by the requirement that the tokens being inserted must parse in isolation.  I'm planning to have it just match nested braces and be done with it.
This implies that we enter a different parsing mode.  I had considered the possibility
that this would mean substituting a different automaton, but then I realized that I should
be able to achieve what I want by, in effect, just gluing together the two automata in
question into one larger one using productions like "TB_START <- TB_START TOKEN" and
eating up all the tokens I see.  I ran into trouble with this approach in that a parser
presented with an instance of TOKEN will always prefer to advance along the automaton
in the direction of the token's symbol, rather than TOKEN itself.

Long story short, I realized what I want is a bit like a regex's "." character, which
matches anything, which in Ellesmere is ANY.  Every symbol ever created should
automatically be a subtype of ANY to make this work.  Following this reasoning leads to
the conclusion that the symbol table must be aware of the inheritance hierarchy so it
can add to it as new symbols are added.  It also led to the idea that the IR nodes should
be one and the same as the symbol token objects, though I'm not yet sure whether this
idea is feasible or desirable.

This line of thought also led me to realize I have a problem if the inheritance hierarchy
changes after the automaton is created (which would happen quite regularly if every new
symbol encountered had to inherit ANY).  I'm not planning to make the heroic effort to
amend the automaton to account for arbitrary inheritance changes, because a new inheritance
relation between two symbols that already existed when the automaton was created would be
particularly disruptive.  However, if I restrict myself to inheritance changes where the
subtype is newer than the automaton, and it has a single supertype T that is already
handled by the automaton, then I should be able to keep using that automaton by treating
the new symbol as a T. (I think I may even be able to augment the automaton to add new
edges wherever there's a T edge.  That would keep the parser simple because it can remain
ignorant of the inheritance hierarchy.)

Anyway, in the mean time, I have implemented the following bug fixes and cleanup items that
I developed during this phase of things.  The plan is then to refactor to make simpler
TOKEN_BLOCK parsing easier to implement, then implement it.

- Fix ts_advance so it doesn't skip an extra token from theLexTokenStream
- Add all initial productions to productionMap -- not just the shallowest productions.  (This doesn't matter as long as the initial grammar is not nested.)
- Add STRING as a subtype of ANY
- Add buffering on diagnostic files
- Fix ob_sendTo to count chars sent
- Fix mh_reallocInPlace to refuse to grow an object if it would extend past the end of the MemoryHunk
- Make make pn_copy call pra_nextElement before it privatizes things, because pra_nextElement can cause the pra to grow (ie move), thereby making privatized pointers obsolete
- Add ANY as a built-in symbol
- Minor RAS stuff
- Add print.el program I ultimately want to get working



git-svn-id: file:///svn/ellesmere@152 8a67d1f4-d520-4d41-b900-46b8824da3b7