DOESN'T WORK
I took a bit of a hiatus starting October 8 to work on the problem of parsing in the
presence of inheritance.  I was unsuccessful in finding a clean way to do it, and now
I'm planning simply to apply some ad hoc techniques to reduce the size of the generated
automaton, hopefully making it reasonably compact in most cases.

I'm also planning to improve the testing framework and diagnostic output to help me work
through these automaton compression problems.  Most importantly, I'd like to improve the
regression testing using .t.c files instead of #ifdefs.

Returning to the project, I found a number of changes to the files.  I don't remember the
reasoning behind some of them, since it has been three months since I made the changes,
but I'll do my best to describe them here.

- Added bv_dup
- Made conflictLog optional.  I think I was seeing gobs of uninteresting conflicts at
  the time, so I just wanted the ability to ignore them.
- Some log file cleanup
- Fixed a bug where an ItemSet pointer could become stale when the underlying array got reallocated
- Reordered a bunch of the code in parser.c.  I forget the exact reason for this.  I think
  I was experimenting with calling some of the functions from new places, so the definitions
  had to be moved above those places.  I vaguely remember it got a bit out of hand, but I'm
  committing it anyway just so I don't lose the record of what I did.
- Fixed a bunch of bugs in pg_computeLR0StateNodes.  Got rid of the notion of itemsLeft for
  a reason I don't recall.
- Added propagateDownward, propagateUpward, and propagateBoth, and used them in place of sst_augment
  so I could control which bitvectors are propagated in which direction at which points in
  the parser generation process.  This was after attempting to augment sste_propagate to do
  the "right" bitvectors for each direction, finding that inadequate, and ultimately getting
  rid of sste_propagate.  PropagateBoth now has the temporary-table logic that used to be
  in sst_augment.



git-svn-id: file:///svn/ellesmere@163 8a67d1f4-d520-4d41-b900-46b8824da3b7