Got the meta-interpreter to fail on a primitive.  Yay!
I finally got this to the point that it fails trying to run a primitive
procedure on behalf of the executing program.  Below are a few of the things I
did.

Create items in reverse order to favour canonicalizing to inner productions
instead of inherited ones.

Raise some exceptions to make things fail more promptly so I can tell what went
wrong: Missed_Action_Word, Unexpected_token, and one for the fact that
do_action_PRIMITIVE is not yet implemented.

Fix a few cases in the meta-interpreter where I had still been using th instead
of th.activation.

Create wrap_procedure that inserts a level of meta-interpretation.  At some
point I'll be able to try adding multiple levels to test the robustness of
meta-interpretation.

A hack to treat RRCONFLICT as REDUCE0.  Really, the logic to detect RRCONFLICT
is busted.  It counts how many items there are in an item set, when it really
should be scanning for differing reduce actions in that state.  (Also I want to
deal with REDUCE1 at some point, to implement lookahead, so that may play a
role here too.)

Problem: building the automaton using Ellesmere is now taking almost 20
minutes.  I REALLY need a better way to deal with inheritance than just
generating the full combinatorial explosion of productions.