Got primitives working
The remaining problem is in parsing something like this:

	some_object 'EOF' 'get'

The parser generator doesn't distinguish between the literal symbol 'EOF' and
an object whose tag is 'EOF', and the parser implicitly resolves the conflict
in favour of treating 'EOF' as 'EOF' rather than a symbol:

	( state obj probe ) next_state2_TAKE_FAILED
		state obj tag get

	( state obj probe ) next_state2_STATE
		probe

	( state obj ) next_state
			state
			obj
				state
				obj
				TAKE_FAILED
			take
		next_state2

This implementation of next_state2 always looks for an EOF edge before trying
the tag of EOF (which would be SYMBOL).  The natural fix here would be to try
SYMBOL first, but it is not at all obvious that this is any better.

The better answer would be to reserve judgement until a reduce decision is
actually necessary.  That would require help from the parser generator, but
again, it's not clear that this is much better.  For one thing, there would be
even more of an explosion of possibilities if nearly every token (or at least
every token not assigned a name in the production) could also be interpreted as
a symbol.  For another thing, it's not at all clear that this ambiguity between
whether something is a literal or a symbol is even a desirable or useful
property of a language.  I'm not sure it's worth supporting.

This difficulty seems to play into a larger them of the difficulty Sheppard has
with quoting: if a name is bound in the current environment, Sheppard simply
offers no way to express that name as a symbol.  It always gets looked up in
the environment.  I have a feeling this problem may be related, but I'm not yet
sure how.  On the one hand, I want it to be possible for the productions to
indicate that quoting is implicit.  On the other hand, I want programs to be
able to do (or undo) quoting explicitly if desired.

One possible way to resolve this is if the parser generator treated unnamed
tokens literally (as keywords) and named tokens as object tags.  I haven't
yet thought through whether this is workable, but it seems kind of neat and
simple.