Prepend colons for automaton edges for object tags
Fails in the same spot as before.  Not yet sure why.

...But differentiating edges followed due to object tags versus edges followed
due to literal symbols is something that has been a long time coming.  It's
something that needed fixing, even if it didn't actually help get through the
EOF problem.

Added @ syntax to the GrammarLine processing, not because the @ syntax is so
great, but because when I made the colon syntax include the colon in the symbol
name, the @ syntax became a natural way NOT to include a colon, which I need
very occasionally.