parse_library can make a LIBRARY from text
It can automatically discover all the symbols in the program and produce the
automaton as well as the action bindings.

Problem: I needed to exclude sharp symbols, or else the automaton will follow
those edges instead of :SYMBOL.  I need to think about why this was necessary
and whether this is really kosher, or whether it's a hack.