parse_library breaks words at whitespace, and edge_symbols
Also, get rid of all_symbols and the sharp hack.  The only symbols that need
edges are ones that can be interpreted as something other than :SYMBOL, and
that only includes symbols that could cause a reduce, plus the object types.
I've added a lame square-bracket syntax to declare all the object types.

This thing would be much simplified if there were inheritance, or at least
default productions.  I'm holding off on doing that just because it complicates
the interpreter, which is already complicated enough, but perhaps it would
simplify the parser generator enough to make it worthwhile, especially once the
parser generator is written in Sheppard code.