NOT WORKING.  Bug fixes and cleanup.
The big thing was that I realized the "nop" trick for eval (used to make sure
the digression doesn't disappear before the string is evaluated) is not really
suitable for prefix programs.  The "nop" tokens just build up at the end of the
program, and they keep a lot of digressions alive that could otherwise be
finished.  I changed eval and exec so they simply take the desired environment
as an argument, rather than grabbing it implicitly.  This way, the correct
environment is captured at the time the current_environment primitive runs,
at which time the right environment is still on the stack.

Also, a number of small things:

- Bug fix: Moved argument splitting inside the factory class so it only happens
  during argument processing.

- Cleanup: Removed calculation of edge_symbols because that's no longer used.

- Differentiated between ":" syntax and "@" syntax so I could specify that I
  want a literal symbol TAKE_FAILED rather than an object of that type.