Add comments and rename NOTHING to DIAL_TONE
Also tons of cleanup
- Remove the hash_test thing.  At the moment, I think my story for hash
tables will be to turn hashable keys into symbols (ie. strings) and just
map them using the Sheppard object model
- Stop dorking dial_tone.bindings.  It has been ages since we used this
to map actions, so there's no need to do this anymore, and it was just
confusing and weird.
- Simplify the argument list to polymorphic_automaton.  I'm hoping to
rewrite this in Sheppard pretty soon, so I need it as simple as
possible.