Replace PRIMITIVE with more specific action types
PRIMITIVE was flawed because it relied on a non-Sheppard object (the
Python function object) being stored in a Sheppard object graph.  I
didn't think too hard about this until I started to add Shogun, at which
point it was impossible to serialize PRIMITIVEs.

I've replaced PRIMITIVE with several classes that each does its own
thing: PYTHON_EVAL, PYTHON_EXEC, and CURRENT_THREAD to handle that
oddball built-in feature that seems to defy any kind of rational design.
The implementations of these are fairly straightforwardly implemented in
the interpreter.  (The meta-interpreter implementations can be
mind-bending, though.  If it's obvious to you that perform on a
PYTHON_EVAL object should be implemented using python_exec, then you are
smarter than was the first time through.)

As for nomenclature, I suddenly decided that you "perform" actions, and
you "process" states.  I'm not super happy with this terminology, but
there you have it.

I also added the capability to perform a round-trip deshogun/shogun
test.  Had to use ast.literal_eval on the strings.  The "findall"
pattern that feeds words to the deshogun process is far from robust.