Reduce confusion over the various environments
In the process, I noticed that LIBRARY doesn't need an environment, and
execute does not need to be given a "local scope" environment because
it's obvious how to construct one.

The global scope objects are still doing double-duty as the action
bindings.  Seems like that ought to be cleaner somehow.