Add DIALECT and FALLBACK objects
Previously, a "dialect" was just a pointer to the initial state of the
parsing automaton.  I'm creating a proper "dialect" object which will
also contain the "fallback relation" (i.e. the implementation side of
the inheritance relation).

I'm not really sure a LIBRARY and a DIALECT are two different things.
Right now, a library is little more than a pointer to the initial state
of a parsing automaton.  It feels like a LIBRARY should also point to
its action bindings, because otherwise the routines defined in the
library are not reachable from the LIBRARY object.  On the other hand, a
DIALECT feels like a way of parsing, which would comprise an automaton
and a fallback relation, and probably not the action bindings.