Some action_bindings cleanup
There's more work to be done, but for now:

Define current_thread in prologue_text.  That was the last of the
definitions from define_builtins, so I was able to get rid of that,
which then meant that parse_library no longer needs to build upon an
existing action_bindings object.  Now it can make its own fresh bindings
object, and that is now part of the returned LIBRARY.

This highlighted the fact that we were pretty loosey-goosey about
global_scope and action_bindings.  I've now separated the two
completely.  For now I've stuffed action_bindings into the PROCEDURE
object, though I'm not sure it will end up there.

In the process of doing that, I discovered that
ACTIVATION.action_bindings was actually an ENVIRONMENT, and I was
looking up actions using the bound() funciton.  I don't think there's a
need for that, so now it's just a BINDINGS.

Also, get rid of global var current_thread because it was confusing when
we also have a Sheppard builtin with the same name.  Turns out it was
only used in one function, and that function's caller had the thread
available anyway.