Build automaton from grammar (not action_bindings)
This "grammar" is still embryonic because the PRODUCTIONs have no lhs,
and the grammar has no goal symbol.  To build a proper parser generator,
I'm going to have to reckon with the return types of actions, and that
leads to dealing with inheritance, and also with how to deal with
actions that don't return anything.  For now, I'm still using the
function "build_keyword_based_automaton" (formerly called just
"polymorphic_automaton") which assumes the first symbol in each
production is a keyword that signals the start of that production, and
that all typing is dynamic, so it doesn't even consult lhs or goal
symbols.

PRODUCTION.rhs is also backward.  (It's called rhs_stack to highlight
this.)  This is not meant to be permanent; it's only because the parser
generator is currently structured the way it was when it was using the
formal_arg_types stack field from the actions, which I have now removed.

Other stuff:
- A feeble attempt to support anonymous symbols in shogun.  Doesn't
really work, because it comes back to that whole issue of object
identities that I need to deal with for true, false, dial_tone, epsilon,
etc. (and which I just bypassed for null by hard-coding that to be
object zero).