Rudimentary parsing of productions
Not yet actually using them to build the automaton yet.  Also, no
handing of associativity or precedence, which will be important
immediately because the main thing I want this for is arithmetic.

The parser is currently creating a KEYWORD_GRAMMAR containing both
KEYWORD_PATTERNs and PRODUCTIONs.  This it a temporary state for
debugging.

A few more odds and ends:
- Fix ANONs without pseudonym so they don't end up with a pseudonum of
None (that's right, the Python None, not the Sheppard null)
- Take names out of MACRO, PYTHON_EVAL, etc.  The ones in productions
won't have names.  For actions coming from keyword patterns, we can
stuff the keyword in there just as an FYI, but we shouldn't be relying
on it for correctness in any way.
- Added some early prototype code to specify arithmetic syntax, and
verified that it parses.
- Renamed the vague PATTERN into KEYWORD_PATTERN
- Removed DEFINITION.  parse_definition now creates KEYWORD_PATTERNs and
PRODUCTIONs directly as appropriate.