Generators to iterate states and state transitions
Also an attempt to combine fib and arithmetic to allow infix + between
two fib calls. That led to a shift-reduce conflict that I don't really
feel like debugging.

Instead, I think I'm going to switch to implementing the eta-cetera
representation of NFAs, and using that to implement SLR parser
generation with symbol inheritance.  (We'll see... my plans seem to
change on a dime.)